Update project to tika-1.20
