+++ Gemfile
@@ -1,0 +1,4 @@
+source "https://rubygems.org"
+
+# Specify your gem's dependencies in rtika.gemspec
+gemspec
