+++ lib/rtika/version.rb
@@ -1,0 +1,3 @@
+module RTika
+  VERSION = "0.3.0"
+end
