+++ rtika.gemspec
@@ -1,55 +1,29 @@
-# Generated by jeweler
-# DO NOT EDIT THIS FILE DIRECTLY
-# Instead, edit Jeweler::Tasks in Rakefile, and run the gemspec command
-# -*- encoding: utf-8 -*-
 
-Gem::Specification.new do |s|
-  s.name = %q{rtika}
-  s.version = "0.3.0"
+lib = File.expand_path("../lib", __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require "rtika/version"
 
-  s.required_rubygems_version = Gem::Requirement.new(">= 0") if s.respond_to? :required_rubygems_version=
-  s.authors = ["Pradeep Elankumaran"]
-  s.date = %q{2010-11-09}
-  s.description = %q{rTika is a JRuby wrapper around the Apache Tika content extraction library}
-  s.email = %q{pradeepe@gmail.com}
-  s.extra_rdoc_files = [
-    "LICENSE",
-     "README.rdoc"
-  ]
-  s.files = [
-    ".document",
-     ".gitignore",
-     "LICENSE",
-     "README.rdoc",
-     "Rakefile",
-     "VERSION",
-     "lib/rtika.rb",
-     "lib/tika-app-0.9-SNAPSHOT.jar",
-     "lib/tika-bundle-0.9-SNAPSHOT.jar",
-     "lib/tika-core-0.9-SNAPSHOT.jar",
-     "lib/tika-parsers-0.9-SNAPSHOT.jar",
-     "rtika.gemspec",
-     "test/helper.rb",
-     "test/test_rtika.rb"
-  ]
-  s.homepage = %q{http://github.com/skyfallsin/rtika}
-  s.rdoc_options = ["--charset=UTF-8"]
-  s.require_paths = ["lib"]
-  s.rubygems_version = %q{1.3.6}
-  s.summary = %q{A JRuby wrapper around the Apache Tika library}
-  s.test_files = [
-    "test/helper.rb",
-     "test/test_rtika.rb"
-  ]
+Gem::Specification.new do |spec|
+  spec.name          = "rtika"
+  spec.version       = RTika::VERSION
+  spec.authors       = ["Pradeep Elankumaran"]
+  spec.email         = ["pradeepe@gmail.com"]
 
-  if s.respond_to? :specification_version then
-    current_version = Gem::Specification::CURRENT_SPECIFICATION_VERSION
-    s.specification_version = 3
+  spec.summary       = %q{A JRuby wrapper around the Apache Tika library.}
+  spec.description   = %q{rTika is a JRuby wrapper around the Apache Tika content extraction library.}
+  spec.homepage      = "http://github.com/skyfallsin/rtika"
+  spec.license       = "MIT"
 
-    if Gem::Version.new(Gem::RubyGemsVersion) >= Gem::Version.new('1.2.0') then
-    else
-    end
-  else
+  # Specify which files should be added to the gem when it is released.
+  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
+  spec.files         = Dir.chdir(File.expand_path('..', __FILE__)) do
+    `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
   end
-end
+  spec.bindir        = "exe"
+  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.require_paths = ["lib"]
 
+  spec.add_development_dependency "bundler", "~> 2.0"
+  spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_development_dependency "rspec", "~> 3.0"
+end
