+++ spec/rb_files/base_file.rb
@@ -1,0 +1,21 @@
+require 'pathname'
+
+module RbFiles
+  class BaseFile
+    def fixture_dir
+      @fixture_dir ||= Pathname.new File.expand_path('../fixtures', __dir__)
+    end
+
+    def name
+      ""
+    end
+
+    def path
+      fixture_dir.join(name)
+    end
+
+    def pages
+      []
+    end
+  end
+end
