+++ spec/rtika_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe RTika do
+  it "has a version number" do
+    expect(RTika::VERSION).not_to be nil
+  end
+end
