--- test/test_rtika.rb
@@ -1,7 +1,0 @@
-require 'helper'
-
-class TestRtika < Test::Unit::TestCase
-  should "probably rename this file and start testing for real" do
-    flunk "hey buddy, you should probably rename this file and start testing for real"
-  end
-end
